emit index on inserted segment

+++ trunk/CLAM/src/Visualization/QTVM/Controllers/PlotController.hxx
@@ -109,7 +109,7 @@ namespace CLAM
 			// propagated signals from SegmentEditor
 			void segmentOnsetChanged(unsigned,double);
 			void segmentOffsetChanged(unsigned,double);
-			void segmentInserted(double);
+			void segmentInserted(unsigned);
 			void segmentDeleted(unsigned);
 			void currentSegmentChanged(unsigned);
